I10-577 InputTextComposite throws Integer to Double cast exception
- remove the explicit Casting to Double.

+++ uk.ac.gda.client/src/gda/rcp/views/InputTextComposite.java
@@ -325,7 +325,7 @@ public class InputTextComposite extends Composite {
 				if (isTextInput()) {
 					positionText.setText(((ScannablePositionChangeEvent)arg).newPosition.toString());
 				} else {
-					positionText.setText(String.format(scannable.getOutputFormat()[0], (Double)((ScannablePositionChangeEvent) arg).newPosition));
+					positionText.setText(String.format(scannable.getOutputFormat()[0], ((ScannablePositionChangeEvent) arg).newPosition));
 				}
 			});
 			} else if (arg.getClass().isArray()) {
